JavaScript resolving dependencies + issue fix with Seeds
