+++ app/models/booking.rb
@@ -1,6 +1,6 @@
 class Booking < ApplicationRecord
   belongs_to :user
   belongs_to :activity
-  has_one :review
+  has_one :review, dependent: :destroy
 
 end
