+++ app/models/category.rb
@@ -1,4 +1,4 @@
 class Category < ApplicationRecord
-  has_many :activities
+  has_many :activities, dependent: :destroy
 
 end
