+++ app/models/user.rb
@@ -3,8 +3,9 @@ class User < ApplicationRecord
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :trackable, :validatable
-  has_many :activities
-  has_many :bookings
+  has_many :activities, dependent: :destroy
+  has_many :bookings, dependent: :destroy
   has_many :followers, class_name: 'Following', foreign_key: 'follower_id'
   has_many :followees, class_name: 'Following', foreign_key: 'followee_id'
+
 end
