+++ app/views/activities/filters.html.erb
@@ -7,3 +7,9 @@
 
 
 <% end %>
+
+<%= content_for :after_js do %>
+<script>
+  window.app.loadSlides()
+</script>
+<% end %>
