+++ app/views/layouts/application.html.erb
@@ -24,6 +24,8 @@
       <%= yield %>
       <%= javascript_include_tag 'application' %>
       <%= javascript_pack_tag 'application' %>
+
+      <%= yield :after_js %>
     </div>
   </body>
 </html>
