+++ db/seeds.rb
@@ -5,7 +5,7 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
-  Activity.destroy_all
+  Review.destroy_all
   Category.destroy_all
   User.destroy_all
